Ensure 'yum install' indeed succeeded.
No point in continuing tests, if 'yum install' failed for Engine
or Hosted Engine RPMs. This may be caused by missing deps, usually.

+++ common/deploy-scripts/install_he_4.sh
@@ -1,5 +1,10 @@
 fstrim -va
 yum install -y ovirt-hosted-engine-setup ovirt-engine-appliance
+RET_CODE=$?
+if [ ${RET_CODE} -ne 0 ]; then
+    echo "yum install failed with status ${RET_CODE}."
+    exit ${RET_CODE}
+fi
 rm -rf /dev/shm/*.rpm
 fstrim -va
