+++ common/deploy-scripts/install_he_4_additional_host.sh
@@ -1,5 +1,10 @@
 fstrim -va
 yum install -y ovirt-hosted-engine-setup sshpass
+RET_CODE=$?
+if [ ${RET_CODE} -ne 0 ]; then
+    echo "yum install failed with status ${RET_CODE}."
+    exit ${RET_CODE}
+fi
 rm -rf /dev/shm/*.rpm
 fstrim -va
 echo -e "\nDefaults:root !requiretty\n" >> /etc/sudoers
